fix(ARA2): resolve type mismatches in ARA factory serialization
- Change `struct_size` field type from `ARA::ARASize` to `uint64_t` in YaARAFactorySnapshot and YaARAFactoryConfig structs for consistent serialization
- Add explicit `static_cast<uint64_t>()` when assigning factory and config struct sizes during initialization
- Add explicit `static_cast<ARA::ARASize>()` when reconstructing factory struct size from snapshot in AraFactoryProxy
- Remove anonymous namespace wrapper around AraFactoryProxy class definition
- Fixes type conversion warnings and ensures proper size handling across serialization boundaries

+++ src/common/serialization/vst3/plugin/ara-plugin-entry.cpp
@@ -31,7 +31,7 @@ YaARAFactorySnapshot::YaARAFactorySnapshot(
         return;
     }
 
-    struct_size = factory->structSize;
+    struct_size = static_cast<uint64_t>(factory->structSize);
     lowest_supported_api_generation = factory->lowestSupportedApiGeneration;
     highest_supported_api_generation = factory->highestSupportedApiGeneration;
 
@@ -98,7 +98,7 @@ YaARAFactoryConfig::YaARAFactoryConfig(
     }
 
     has_config = true;
-    struct_size = config->structSize;
+    struct_size = static_cast<uint64_t>(config->structSize);
     desired_api_generation = config->desiredApiGeneration;
     has_assert_function = config->assertFunctionAddress != nullptr;
 }
